fix: process handling on windows
Refactor how process handling is done so that we can customise the setup
of exec.Cmd to create a new process group.

This is needed to ensure that when CTRL_BREAK_EVENT is sent it only
effects our helper process and not the main test which was causing it to
fail.

This also allows the gopro package users to fully customise the process
handling by setting a custom Handler, which may be useful to process or
redirect ffmpeg output.