feat(geo): add two point distance calculation
Add the ability to to calculate the distance between two points using
both haversine and equirectangular which is roughly 5x faster taking
20ns vs 92ns on a i7-7700HQ CPU @ 2.80GHz.

This refactors the previous Matcher into Processor as it now provides
more than just the ability to match.

As units are based on the passed in radius it also removes all mention
of meters from function definitions.

Eliminate the import of geodesic which was only needed for the earth
radius from the WGS84 definition.